chore: covert FaceLogo to ts
