+++ src/containers/unit/Footer/styles/desktop_view/digest_view/index.ts
@@ -23,7 +23,7 @@ export const TopBarInfos = styled.div`
   ${css.flex('align-center')};
 `
 export const MainInfos = styled.div`
-  ${css.flexColumn('justify-start')};
+  ${css.flex('justify-start')};
   margin-bottom: 20px;
   margin-top: 20px;
   margin-bottom: 30px;
